+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/BasePaymentMethodsListFragment.kt
@@ -61,6 +61,13 @@ internal abstract class BasePaymentMethodsListFragment(
         isEditing = savedInstanceState?.getBoolean(IS_EDITING) ?: false
     }
 
+    override fun onResume() {
+        super.onResume()
+
+        sheetViewModel.headerText.value =
+            getString(R.string.stripe_paymentsheet_select_payment_method)
+    }
+
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         inflater.inflate(R.menu.paymentsheet_payment_methods_list, menu)
         // Menu is created after view state is restored, so we need to update the title here
@@ -127,6 +134,7 @@ internal abstract class BasePaymentMethodsListFragment(
         adapter.setItems(
             config,
             sheetViewModel.paymentMethods.value.orEmpty(),
+            sheetViewModel is PaymentOptionsViewModel,
             sheetViewModel.selection.value
         )