+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentOptionsViewModelTest.kt
@@ -64,7 +64,8 @@ internal class PaymentOptionsViewModelTest {
         logger = Logger.noop(),
         injectorKey = DUMMY_INJECTOR_KEY,
         resourceRepository = resourceRepository,
-        savedStateHandle = SavedStateHandle()
+        savedStateHandle = SavedStateHandle(),
+        linkPaymentLauncherFactory = mock()
     )
 
     @Test
@@ -141,7 +142,8 @@ internal class PaymentOptionsViewModelTest {
             logger = Logger.noop(),
             injectorKey = DUMMY_INJECTOR_KEY,
             resourceRepository = resourceRepository,
-            savedStateHandle = SavedStateHandle()
+            savedStateHandle = SavedStateHandle(),
+            linkPaymentLauncherFactory = mock()
         )
 
         var transitionTarget: BaseSheetViewModel.Event<TransitionTarget?>? = null
@@ -172,7 +174,8 @@ internal class PaymentOptionsViewModelTest {
             logger = Logger.noop(),
             injectorKey = DUMMY_INJECTOR_KEY,
             resourceRepository = resourceRepository,
-            savedStateHandle = SavedStateHandle()
+            savedStateHandle = SavedStateHandle(),
+            linkPaymentLauncherFactory = mock()
         )
 
         val transitionTarget = mutableListOf<BaseSheetViewModel.Event<TransitionTarget?>>()
@@ -203,7 +206,8 @@ internal class PaymentOptionsViewModelTest {
             logger = Logger.noop(),
             injectorKey = DUMMY_INJECTOR_KEY,
             resourceRepository = resourceRepository,
-            savedStateHandle = SavedStateHandle()
+            savedStateHandle = SavedStateHandle(),
+            linkPaymentLauncherFactory = mock()
         )
 
         val transitionTarget = mutableListOf<BaseSheetViewModel.Event<TransitionTarget?>>()
@@ -234,7 +238,8 @@ internal class PaymentOptionsViewModelTest {
             logger = Logger.noop(),
             injectorKey = DUMMY_INJECTOR_KEY,
             resourceRepository = resourceRepository,
-            savedStateHandle = SavedStateHandle()
+            savedStateHandle = SavedStateHandle(),
+            linkPaymentLauncherFactory = mock()
         )
 
         viewModel.removePaymentMethod(cards[1])